Start to trim the inventory API
